Share `secp256k1` code
