+++ crates/interpreter/Cargo.toml
@@ -11,7 +11,6 @@ readme = "../../README.md"
 
 [dependencies]
 revm-primitives = { path = "../primitives", version = "1.3.0", default-features = false }
-secp256k1 = { version = "0.28.0", default-features = false, features = [ "alloc", "recovery", ] }
 
 # optional
 serde = { version = "1.0", features = ["derive", "rc"], optional = true }
