+++ crates/primitives/src/lib.rs
@@ -17,6 +17,7 @@ pub mod kzg;
 mod log;
 pub mod precompile;
 pub mod result;
+pub mod secp256k1;
 pub mod specification;
 pub mod state;
 pub mod utilities;
